+++ packages/react/__tests__/snapshots.test.js
@@ -1,5 +1,0 @@
-import initStoryshots from '@storybook/addon-storyshots'
-import registerRequireContextHook from 'babel-plugin-require-context-hook/register'
-
-registerRequireContextHook()
-initStoryshots()