+++ packages/react/src/layout/Modal.js
@@ -3,7 +3,7 @@ import * as F from 'futil-js'
 import { observer } from 'mobx-react'
 import Portal from './Portal'
 
-let Modal = ({ isOpen, children, style = {} }) => (
+let Modal = ({ isOpen, children, style = {}, className = '' }) => (
   <Portal>
     {F.view(isOpen) && (
       <div
@@ -23,7 +23,7 @@ let Modal = ({ isOpen, children, style = {} }) => (
           ...style,
         }}
         onClick={F.off(isOpen)}
-        className="default-modal-bg"
+        className={`default-modal-bg ${className}`}
       >
         <div
           style={{